requirements.txt
Django==1.4
PIL==1.1.7
Pygments==1.6
South==0.8.1
Whoosh==2.4.1
amqp==1.2.0
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.31
celery==3.0.21
celery-with-redis==3.0
django-appconf==0.6
django-attachments==0.3.1
django-avatar==1.0.5
django-celery==3.0.17
django-debug-toolbar==0.9.4
django-extensions==1.1.1
django-guardian==1.1.1
django-haystack==2.0.0
django-kombu==0.9.4
django-picklefield==0.3.0
django-registration==1.0
django-staticfiles==1.2.1
django-tagging==0.3.1
ipdb==0.7
ipython==0.13.2
kombu==2.5.12
pygooglechart==0.3.0
pyparsing==1.5.7
python-dateutil==2.1
pytz==2013b
redis==2.7.6
six==1.3.0
wsgiref==0.1.2
